feat: 評価の平均値手動計算

+++ data/client/review.go
@@ -26,7 +26,7 @@ type ReviewListingResponse struct {
 	ShopName   string                         `json:"shopName"`
 	ServiceID  uint                           `json:"serviceId"`
 	WifiName   string                         `json:"wifiName"`
-	Average    int                            `json:"average"`
+	Average    float32                        `json:"average"`
 	ReviewList []ReviewListingResponseElement `json:"reviewList"`
 	Total      int                            `json:"total"`
 }
